controller, routes and middlewares created

+++ src/controllers/link.controller.js
@@ -3,7 +3,7 @@ import { nanoid } from 'nanoid'
 
 export const createLink = async (req, res) => {
   const { url, description, linky, tags } = req.body
-  const userId = req.user.id
+  const userId = req.userId
   try {
     const shortLink = linky || nanoid(10)
 
@@ -33,7 +33,7 @@ export const createLink = async (req, res) => {
 }
 
 export const getLinks = async (req, res) => {
-  const userId = req.user.id
+  const userId = req.userId
   try {
     const links = await prisma.links.findMany({
       where: { creatorId: userId },
@@ -48,12 +48,21 @@ export const getLinks = async (req, res) => {
 
 export const updateLink = async (req, res) => {
   const { id } = req.params
-  const { url, description, tags } = req.body
+  const { url, description, linky, tags } = req.body
   try {
+    const shortLink = linky || nanoid(10)
+
+    const existingLink = await prisma.links.findUnique({
+      where: { linky: shortLink }
+    })
+
+    if (existingLink) return res.status(400).json({ error: 'Link already exists' })
+
     const updatedLink = await prisma.links.update({
       where: { id: parseInt(id) },
       data: {
         url,
+        linky: shortLink,
         description,
         tags: {
           deleteMany: {},
@@ -76,5 +85,21 @@ export const deleteLink = async (req, res) => {
   } catch (error) {
     console.error(error)
     res.status(500).json({ error: 'Error deleting link' })
+  }
+}
+
+export const redirectLink = async (req, res) => {
+  const { linky } = req.params
+  try {
+    const link = await prisma.links.findUnique({
+      where: { linky }
+    })
+
+    if (!link) return res.status(404).json({ error: 'Link not found' })
+
+    res.redirect(link.url)
+  } catch (error) {
+    console.error(error)
+    res.status(500).json({ error: 'Error redirecting link' })
   }
 }
